+++ lib/roadie/asset_provider.rb
@@ -5,7 +5,7 @@ module Roadie
   # It helps you by declaring {#find_stylesheet!} in the terms of #find_stylesheet in your own class.
   module AssetProvider
     def find_stylesheet!(name)
-      find_stylesheet(name) or raise CssNotFound, name
+      find_stylesheet(name) or raise CssNotFound.new(name, nil, self)
     end
   end
 end